requirements.txt
hvac==0.10.5
ansible==2.9.10
meraki-sdk==1.5.0
pan-python==0.16.0
pandevice==0.14.0
paramiko==2.7.2
ncclient==0.6.9
